% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{add.pointflux}
\alias{add.pointflux}
\title{Add a point flux}
\usage{
add.pointflux(model, at, rate, name)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FLOW2D} (as generated by
\code{newFLOW2D} or \code{newAFLOW2D}) or \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}

\item{at}{a vector containing two number being the x and y coordinate of the
place in the domain where to add the point flux}

\item{rate}{is the magnitue of the flux. Its physical dimension is that of the internal
flux multiplied by length. It can be given by either
\itemize{
\item a number,
\item a string giving a variable name (at every evalution later the value of that variable
at that time will be used)
\item or a function of  one argument (the state s) that
 returns the rate of the point flux
as a function of the state argument}}

\item{name}{optional name for the flux}
}
\description{
Adds an external point flux to an existing model
}
\examples{
invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")
# add a rate and give a name
add.pointflux(M,c(3,3),-1,"p1")
# add a string as a reference to a variable
p2 = -2
add.pointflux(M,c(4,4),"p2")
# add a function
pfunc = function(s){return(-2+s)}
add.pointflux(M,c(5,5),pfunc)
summary(M)
}
