% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{add.spatialflux}
\alias{add.spatialflux}
\title{Add spatial flux}
\usage{
add.spatialflux(model, rate, name = "")
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FLOW2D} (as generated by
\code{newFLOW2D} or \code{newAFLOW2D}) or \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}

\item{rate}{the intensity of the flux. Its physical dimension is that of the internal
flux divided by length. It can be given by either
\itemize{
\item  a number,
\item a string giving a variable name (at every evalution later the value of that variable
at that time will be used)
\item or a function  of space (x,y) and state (s) returning the intensity
of the spatial flux at that point with that state value.}}

\item{name}{an optional name for the flux}
}
\description{
Adds an external spatial flux to an existing model
}
\examples{
invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")
# add a constant rate and give name
add.spatialflux(M,0.1,"s1")
# add a string as a reference to a variable
F=0.1
add.spatialflux(M,"F")
# add a function
sfunc = function(x,y,s){return(-2+sin(x+y)+s)}
add.spatialflux(M,sfunc)
summary(M)
}
