% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{calc.misfits}
\alias{calc.misfits}
\title{Calculate misfits}
\usage{
calc.misfits(model)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}
}
\value{
returns a vector with for each node the calculated misfit
}
\description{
Calculates the misfits, i.e. how much the balances calculated for each node
with the current state values differ from zero. If the states are close to the
solution of the flow problem, these numbers should be small.
}
\examples{
invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")
set.BC.fixedstate(M,"NE",2)
set.BC.fixedflux(M,1,0.4)
set.discretisation(M,nodes=list(type="hexagonal",side=1),method="FV")
hist(calc.misfits(M))
}
