% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{calc.RMSM}
\alias{calc.RMSM}
\title{Calculate RMSM}
\usage{
calc.RMSM(model)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}
}
\value{
returns the Root Mean Square Misfit
}
\description{
Calculates the root mean square error of  the misfits in the active nodes
(=nodes where no given state boundary condition).
If the states are close to the
solution of the flow problem, this number should be small.
}
\examples{
invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")
set.BC.fixedstate(M,"NE",2)
set.BC.fixedflux(M,1,0.4)
set.discretisation(M,nodes=list(type="hexagonal",side=1),method="FV")
calc.RMSM(M)
}
\seealso{
\code{calc.misfits}, \code{calc.MAM}
}
