% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{copy.model}
\alias{copy.model}
\title{Create a copy of a model}
\usage{
copy.model(model)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FLOW2D} (as generated by
\code{newFLOW2D} or \code{newAFLOW2D} or \code{newAFLOW2D}) or \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}
}
\value{
a copy of the same class
}
\description{
Makes a (deep)copy of an two dimensional flow object (the
usual assignment operators "=" and "<-" only generate a new name to an
existing model)
}
\examples{
invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M0 = newFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")
M1 = copy.model(M0)
set.name(M1,"a copy")
summary(M0)
summary(M1)
}
