% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{dataframe.balance}
\alias{dataframe.balance}
\title{Get balance overview}
\usage{
dataframe.balance(model, region = numeric(0))
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}

\item{region}{a vector of nodes for which the balance is calculated, if empty
(which is default) the region will the whole domain}
}
\value{
returns a data frame consisting of four columns: first column contains
the name of the flux, the second all contributions of that flux that result
in a inflow into the domain, the third the contributions resulting in an outflow
from the region, the fourth the net contribution to the balance of the region.
The last row of the data frame contains the sum over the columns.
}
\description{
Returns a data frame of the balance terms for a region defined by
a subset of notes of the
domain  as calculated with the current state values.
}
\examples{
invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")
set.discretisation(M,nodes=list(type="hexagonal",side=1),method="FV")
add.spatialflux(M,1)
add.pointflux(M,c(5,5),-30,"pf")
set.BC.fixedstate(M,"NE",2)
set.BC.fixedflux(M,1,0.4)
solve.steps(M)
plot(M,fill="state",nodenumbersize=1)
dataframe.balance(M,c(14,15,19,20))
dataframe.balance(M)
}
