% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{dataframe.boundaries}
\alias{dataframe.boundaries}
\title{Get boundary fluxes}
\usage{
dataframe.boundaries(model)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D}, \code{[A]FEquartic2D} or \code{FV2D}
(as generated by \code{set.discretisation})}
}
\value{
Returns a data frame consisting of five columns: first column contains
the nodenumber,  the second the total flux over the boundary in the mass balance
of the node, third column the boundary type, fourth and fifth column the domain segment number
and  the wind direction id of
that segment, as also given by the \code{plot.BC.id} function
}
\description{
Returns a data frame of the fluxes over the boundaries of the
domain  as calculated with the current state values. If the boudary
condition was a fixed state, the flux is calculated such that the
balance in that point is statisfied
}
\examples{
invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")
set.discretisation(M,nodes=list(type="hexagonal",side=1),method="FV")
set.BC.fixedstate(M,"NE",2)
set.BC.fixedflux(M,1,0.4)
solve.steps(M)
dataframe.boundaries(M)
}
