% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{dataframe.externalfluxes}
\alias{dataframe.externalfluxes}
\title{Get current external fluxes}
\usage{
dataframe.externalfluxes(model)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}
}
\value{
returns a data frame consisting of several columns:
\itemize{
\item first and second columns ("x" and "y")
 contain the coordinates of the nodes
\item the following columns contain spatial external fluxes (if present) both
as intensities (column name equals fluxname) in the nodepoint
an as the total intergrated contribution to the balance of that node
(column name is \code{S(fluxname)dxdy})
\item the next columns contain the point flux (if present) contribution to
the balance of the node
(column name is pointflux name);
for FV models this contribution goes completely to the node nearest to the
position of the point flux,
for FE models the contribution is distributed over the nodes of the
element containing the position of the point flux.
\item the last columns contain the line flux (if present) both as
intensities (column name equals fluxname) in the nodepoint and as
total integrated contribution
to the balance of that node (column name is \code{S(lineflux)dl})
}
}
\description{
Returns a data frame of the external fluxes in the nodes as calculated with
the current state values
}
\examples{
  invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")
add.pointflux(M,c(2,3),1,"pf")
sfunc = function(x,y,s){return(-(x+y)/10)}
add.spatialflux(M,sfunc)
set.discretisation(M,nodes=list(type="hexagonal",side=1),method="FE")
head(dataframe.externalfluxes(M))
}
