% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{dataframe.internalfluxes}
\alias{dataframe.internalfluxes}
\title{Get current internal fluxes}
\usage{
dataframe.internalfluxes(model)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}
}
\value{
returns a data frame consisting of several columns giving an overview
of the internal fluxes as calculated with the current states.
\itemize{
\item if the numerical method is of the Finite Volume type the first two columns
give the nodenumbers between which the internal flux is calculated, the second column
gives the flux intensity, and the third column the total flux over the side that
separates the volume of the first node from the volume of the second node
\item if the numerical method is of the Finite Element type, the first three columns
give the nodenumbers of the triangular element,
the columns 4 to 6 give the total fluxes calculated in this triangle,
and column 7,8 give of the total flux vector its x and y component.
}
}
\description{
Returns a data frame of the internal fluxes between nodes as calculated with
the current state values
}
\examples{
 invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")

set.discretisation(M,nodes=list(type="hexagonal",side=1),method="FV")
set.BC.fixedstate(M,"NE",2)
set.BC.fixedflux(M,1,0.4)
control=solve.steps(M)
head(dataframe.internalfluxes(M))
set.discretisation(M,nodes=list(type="hexagonal",side=1),method="FE")
set.BC.fixedstate(M,"NE",2)
set.BC.fixedflux(M,1,0.4)
control = solve.steps(M)
head(dataframe.internalfluxes(M))
}
