% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{dataframe.nodes}
\alias{dataframe.nodes}
\title{Get current nodes}
\usage{
dataframe.nodes(model)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}
}
\value{
returns a data frame consisting of two columns names "x" and "y"
containing the coordinates of the nodes (can if needed be converted into a matrix
by the \code{as.matrix} function)
}
\description{
Returns a data frame with current nodes in the model.
}
\examples{
invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")
set.discretisation(M,nodes=list(type="geomgen",optdist=1),
                   method="FV")
head(dataframe.nodes(M))
}
