% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{dataframe.states}
\alias{dataframe.states}
\title{Get current state values}
\usage{
dataframe.states(model)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}
}
\value{
returns a data frame consisting of three columns: first two columns
(named "x" and "y")
contains the coordinates of the nodes, second column (called "state")
the corresponding state values
}
\description{
Returns a data frame with current state values in the model.
}
\examples{
invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")
set.discretisation(M,nodes=list(type="hexagonal",side=1),method="FV")
add.spatialflux(M,0.1)
set.BC.fixedstate(M,"NE",0)
set.BC.fixedstate(M,"SW",0)
solve.steps(M)
plot(M,fill="state")
dataframe.states(M)
}
