% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{do.initialize}
\alias{do.initialize}
\title{Initialize a discrete model}
\usage{
do.initialize(model, init)
}
\arguments{
\item{model}{a two dimensional flow model of the class  \code{[A]FElinear2D},
 \code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}

\item{init}{can be either be either
\itemize{
\item a number: all the states will be initialized to this constant
\item a vector of numbers, whose values (in the order of the nodes) will
be asigned to the states
\item a function of two arguments (x,y); for each node the state value is
calculated by applying this function to its nodal coordinates
}}
}
\description{
With the help of the argument \code{init} all the state values
for the active nodes (i.e. nodes which where not given a fixed state
boundary condition) the model receive a value. As in general the solution method is
iterative, a good initialization may be important.
}
\examples{
invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")
add.spatialflux(M,0.1)
set.BC.fixedstate(M,"NE",2)
set.BC.fixedstate(M,"SW",1)
set.discretisation(M,nodes=list(type="geomgen",optdist=2),method="FV")
do.initialize(M,1.5)
plot(M,fill="state")
}
