% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstructurednodes.R
\name{func.dist.to.line}
\alias{func.dist.to.line}
\title{A distance to a line function constructor}
\usage{
func.dist.to.line(line, width, atvalue, farvalue, vectorarg = TRUE)
}
\arguments{
\item{line}{a matrix of two columns giving the coordinates of the line}

\item{width}{the distance from the line from which the far value 
will be returned}

\item{atvalue}{the value returned at the line}

\item{farvalue}{the value returned further than width from the line}

\item{vectorarg}{(default \code{TRUE}): if TRUE the argument of the returned 
function should be a vector (\code{c(x,y)}), if FALSE the returned function
needs two arguments (\code{x} and \code{y})}
}
\value{
a function specified above
}
\description{
A constructor of a function that gives one value 
on a specified line, another value far from the given line and a smooth 
transition in between. Such a function can be usefull for instance in 
\code{add.spatialflux} and \code{set.discretisation}.
}
\examples{
line = matrix(c(2,2,3,5,8,6),ncol=2,byrow=TRUE)
foo1 = func.dist.to.line(line,width=4,atvalue=4,farvalue=1)
foo1(c(2,3))
foo2= func.dist.to.line(line,width=4,atvalue=4,farvalue=1,vectorarg=FALSE)
foo2(2,3)
xforplot = seq(0,10,length=60)
yforplot = seq(0,10,length=60)
zforplot = outer(xforplot,yforplot,foo2)
surface3d(xforplot,yforplot,zforplot,col="red",alpha=0.8)
}
