% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstructurednodes.R
\name{func.dist.to.point}
\alias{func.dist.to.point}
\title{A distance to point function constructor}
\usage{
func.dist.to.point(point, width, atvalue, farvalue, vectorarg = TRUE)
}
\arguments{
\item{point}{a vector giving the coordinates of the point}

\item{width}{the distance from the point from which the far value 
will be returned}

\item{atvalue}{the value returned at the point}

\item{farvalue}{the value returned further than width from the point}

\item{vectorarg}{(default \code{TRUE}): if TRUE the argument of the returned 
function should be a vector (\code{c(x,y)}), if FALSE the returned function
needs two arguments (\code{x} and \code{y})}
}
\value{
a function specified above
}
\description{
A constructor of a function that gives one value 
at a specified point, another value far from the given point and a smooth 
transition in between. Such a function can be usefull for instance in 
\code{add.spatialflux} and \code{set.discretisation}.
}
\examples{
point=c(4,6)
foo1 = func.dist.to.point(point,width=4,atvalue=4,farvalue=1)
foo1(c(5,5))
foo2 = func.dist.to.point(point,width=4,atvalue=4,farvalue=1,vectorarg=FALSE)
foo2(5,5)
xforplot = seq(0,10,length=60)
yforplot = seq(0,10,length=60)
zforplot = outer(xforplot,yforplot,foo2)
surface3d(xforplot,yforplot,zforplot,col="red",alpha=0.8)
}
