% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstructurednodes.R
\name{func.dist.to.poly}
\alias{func.dist.to.poly}
\title{A distance to a polygon function constructor}
\usage{
func.dist.to.poly(poly, width, invalue, farvalue, vectorarg = TRUE)
}
\arguments{
\item{poly}{a matrix of two columns giving the coordinates of the corners
of the polygon}

\item{width}{the distance from the polygon from which the far value 
will be returned}

\item{invalue}{the value returned inside the polygon}

\item{farvalue}{the value returned further than width from the polygon}

\item{vectorarg}{(default \code{TRUE}): if TRUE the argument of the returned 
function should be a vector (\code{c(x,y)}), if FALSE the returned function
needs two arguments (\code{x} and \code{y})}
}
\value{
a function specified above
}
\description{
A constructor of a function that gives one value 
inside a given polygon, another value far from the given polygon and a smooth 
transition in between. Such a function can be usefull for instance in 
\code{add.spatialflux} and \code{set.discretisation}.
}
\examples{
poly = matrix(c(2.5,2.5,3,5,8,6),ncol=2,byrow=TRUE)
foo1 = func.dist.to.poly(poly,width=4,invalue=4,farvalue=1)
foo1(c(2,3))
foo2 = func.dist.to.poly(poly,width=4,invalue=4,farvalue=1,vectorarg=FALSE)
foo2(2,3)
xforplot = seq(0,10,length=60)
yforplot = seq(0,10,length=60)
zforplot = outer(xforplot,yforplot,foo2)
surface3d(xforplot,yforplot,zforplot,col="red",alpha=0.8)
}
