% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{invdist.fun}
\alias{invdist.fun}
\title{Inverse distance interpolator}
\usage{
invdist.fun(tointerp, power = 2, zname = "z", domain = NULL)
}
\arguments{
\item{tointerp}{a dataframe containing at least a columns called "x" and
"y" which will be used as coordinates of points in the plane}

\item{power}{default 2}

\item{zname}{the name of the column  to be interpolated, default "z"}

\item{domain}{optional; should be a matrix of two column defining a polygon,
outside of this the reulting function wil return a value NA}
}
\value{
a function that for any two arguments returns the interpolated value
}
\description{
constructs a function that performs an inverse distance
interpolation in two  dimensions
}
\examples{
N=100
x = runif(N)
y = runif(N)
W = abs(x-y*y)
ToInt = data.frame(x=x,y=y,W=W)
Npl = 35
xtopl = seq(0,1,length=Npl)
ytopl = seq(0,1,length=Npl)
IntFun = invdist.fun(ToInt,zname="W")
z = outer(xtopl,ytopl,IntFun)
image(xtopl,ytopl,z)
}
\seealso{
\code{lininterpol.fun},\code{nearestneighbour.fun}
}
