% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{jitter.points}
\alias{jitter.points}
\title{jitter the points in a matrix}
\usage{
jitter.points(M, eps = 1e-06, tojitter = c())
}
\arguments{
\item{M}{a matrix of two columns (x and y) who's rows are
points in the plane}

\item{eps}{a measure for the jitter, default 1e-6}

\item{tojitter}{a vector of row numbers corresponding to
the points to be jittered; if empty (=default) all points
are jittered}
}
\description{
jitter (randomly) points to avoid collinearity
}
\examples{
D = matrix(c(0,0,5,0,10,0,10,5,10,10,5,10,0,10,0,5),ncol=2,byrow=TRUE)
print(D)
plot(D)
D = jitter.points(D)
print(D)
points(D,pch=20,col="red")
}
