% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZAFVFE2D.R
\name{newAFLOW2D}
\alias{newAFLOW2D}
\title{Construct a new AFLOW2D model}
\usage{
newAFLOW2D(domain = NULL, systemfluxfunction = NULL, name = "")
}
\arguments{
\item{domain}{a matrix with two columns containing respectively the
x and y coordinates of the cornerpoints that define the polygonal domain
on which the model is to be solved}

\item{systemfluxfunction}{a function of five arugments:
coordinates  x and y defining a point withtin the domain,
the state s at this point and the
gradients dsdx and dsdy of the state at that point; it has to return a
vector of two number that
gives the flux in the x and y direction respecitvely
based on these five arguments}

\item{name}{an optional name of the new object}
}
\value{
an object of the class AFLOW2D
}
\description{
generates a new object of the class AFLOW2D:
the base of a two dimensional flow
}
\examples{
kvalues = data.frame(x=c(3,6),y=c(2,7),k=c(3,10))
kfun = nearestneighbour.fun(kvalues,zname="k")
invgradflux = function(x,y,s,gradx,grady)
{
return(-kfun(x,y)*s*c(gradx,0.2*grady))
}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newAFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")
summary(M)
}
