% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{newFLOW2D}
\alias{newFLOW2D}
\title{Construct a new FLOW2D model}
\usage{
newFLOW2D(domain = NULL, systemfluxfunction = NULL, name = "")
}
\arguments{
\item{domain}{a matrix with two columns containing respectively the
x and y coordinates of the cornerpoints that define the polygonal domain
on which the model is to be solved}

\item{systemfluxfunction}{a function of four arugments:
coordinates  x and y defining a point withtin the domain,
the state s at this point and the
gradient ds of the state at that point; it has to return a number that
gives the flux in the direction of the gradient based on these four arguments}

\item{name}{an optional name of the new object}
}
\value{
an object of the class FLOW2D
}
\description{
generates a new object of the class FLOW2D:
the base of a two dimensional flow
}
\examples{
kvalues = data.frame(x=c(3,6),y=c(2,7),k=c(3,10))
kfun = nearestneighbour.fun(kvalues,zname="k")
invgradflux = function(x,y,s,grads) { return(-kfun(x,y)*s*grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")
summary(M)
}
