% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{plot.BC.id}
\alias{plot.BC.id}
\title{plot BC identifiers}
\usage{
plot.BC.id(model)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FLOW2D} (as generated by
\code{newFLOW2D} or \code{newAFLOW2D}) or \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}
}
\description{
Makes a plot of the domain containing the boundary segment identifiers
(both as winddirections and numbers) of the domain.
These identifiers are needed as argument for the \code{set.BC} functions
}
