% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2Dlineflux.R
\name{rem.lineflux}
\alias{rem.lineflux}
\title{Remove lineflux}
\usage{
rem.lineflux(model, name)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FLOW2D} (as generated by
\code{newFLOW2D} or \code{newAFLOW2D}) or \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}

\item{fluxname}{is the name of the flux to be removed
(use \code{summary} to see the names)}
}
\description{
Removes a line  flux from an existing model
}
\examples{
invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux)
river = matrix(c(3,1,4,6,7,7),ncol=2,byrow=TRUE)
add.lineflux(M,river,rates=list(-2,-1.5),name="river")
summary(M)
rem.lineflux(M,"river")
summary(M)
}
