% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{rem.pointflux}
\alias{rem.pointflux}
\title{Remove point flux}
\usage{
rem.pointflux(model, name)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FLOW2D} (as generated by
\code{newFLOW2D} or \code{newAFLOW2D}) or \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}

\item{fluxname}{is a string giving the name of the flux to be removed
(use \code{summary} to see the names)}
}
\description{
Removes an external point flux from an existing model
}
\examples{
invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux)
add.pointflux(M,at=c(3,3),-1,"p1")
p2 = -2
add.pointflux(M,at=c(4,4),"p2")
pfunc = function(s){return(-2+s)}
add.pointflux(M,at=c(5,5),pfunc)
summary(M)
rem.pointflux(M,"p1")
summary(M)
rem.pointflux(M,"pfunc")
summary(M)
}
