% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{rem.spatialflux}
\alias{rem.spatialflux}
\title{Remove spatial flux}
\usage{
rem.spatialflux(model, name)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FLOW2D} (as generated by
\code{newFLOW2D} or \code{newAFLOW2D}) or \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}

\item{fluxname}{is a string giving the name of the flux to be removed
(use \code{summary} to see the names)}
}
\description{
Removes an external spatial flux from an existing model
}
\examples{
invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux)
add.spatialflux(M,0.1,"s1")
F=0.1
add.spatialflux(M,"F")
sfunc = function(x,y,s){return(-2+sin(x+y)+s)}
add.spatialflux(M,sfunc)
summary(M)
rem.spatialflux(M,"sfunc")
summary(M)
}
