% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{set.BC.fixedflux}
\alias{set.BC.fixedflux}
\title{Set a fixed flux boundary condition}
\usage{
set.BC.fixedflux(model, where, value)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FLOW2D} (as generated by
\code{newFLOW2D} or \code{newAFLOW2D}) or \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}

\item{where}{either a wind direction or a number identifying the side of the domain,
as shown by \code{plot.BC.id}}

\item{value}{either a number or a string giving a variable name
(at every evaluation later the value of that variable
the value of that variable at that time will be used)
for the fluxintensity orthognal to the given side of the boundary,
a postive sign indicating a flux into the domain, a negative sign a flux
leaving the domain}
}
\description{
Set a fixed flux boundary condition to one of the sides  of the polygon
forming the domain of an existing model. A fixed flux of zero is the default
boundary condition set to all boundaries on creation of a flow model.
}
\examples{
invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")
set.discretisation(M,nodes=list(type="hexagonal",side=1),method="FV")
add.spatialflux(M,0.1)
set.BC.fixedstate(M,"NE",0)
set.BC.fixedflux(M,"SW",-0.3)
control = solve.steps(M)
plot(M,fill="state")
}
\seealso{
\code{dataframe.boundaries, plot.BC.id}
}
