% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{set.BC.fluxstate}
\alias{set.BC.fluxstate}
\title{Set a flux-state boundary condition}
\usage{
set.BC.fluxstate(model, where, value)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FLOW2D} (as generated by
\code{newFLOW2D} or \code{newAFLOW2D}) or \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}

\item{where}{either a wind direaction or a number identifying the side of the domain,
as shown by \code{plot.BC.id}}

\item{value}{a function of state given the fluxintensity orthogonal to the given
side, a postive sign indicating a flux into the domain, a negative sign a flux
leaving the domain}
}
\description{
Set a flux-state boundary condition to one of the sides  of the polygon
forming the domain of an existing model
}
\examples{
invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")
set.discretisation(M,nodes=list(type="hexagonal",side=1),method="FV")
add.spatialflux(M,0.1)
set.BC.fixedstate(M,"NE",0)
set.BC.fluxstate(M,"SW",function(s){return(s-0.5)})
control=solve.steps(M)
plot(M,fill="state")
}
\seealso{
\code{dataframe.boundaries, plot.BC.id}
}
