% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{set.isacceptable}
\alias{set.isacceptable}
\title{Set a state-acceptability function}
\usage{
set.isacceptable(model, newisaccfunc)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FLOW2D} (as generated by
\code{newFLOW2D} or \code{newAFLOW2D}) or \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}

\item{newisaccfunc}{a function of three arguments: an x and y giving the place and state return TRUE
if the state is acceptable at that position in the domain and FALSE
if not}
}
\description{
Set a state-acceptability function to an existing model,
to limit the possible values of states to (physcical) acceptable values, as
often implicetely assumed in flux definitions.
}
\examples{
sposgradflux = function(x,y,s,grads) { return(-sqrt(s)*grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=sposgradflux,name="example")
set.isacceptable(M,function(x,y,s){return(s>=0)})
set.discretisation(M,nodes=list(type="hexagonal",side=1),method="FV")
add.spatialflux(M,0.1)
set.BC.fixedstate(M,"NE",0.1)
control = solve.steps(M)
plot(M,fill="state")
}
