% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{set.name}
\alias{set.name}
\title{Set name}
\usage{
set.name(model, newname)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FLOW2D} (as generated by
\code{newFLOW2D} or \code{newAFLOW2D}) or \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}

\item{newname}{the new name for the model}
}
\description{
Set a new name to an existing model
}
