% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{solve.step}
\alias{solve.step}
\title{Do one solution step}
\usage{
solve.step(model)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}
}
\value{
returns a list with two entries: \code{RMSM} which contains the root mean
square of the misfits and \code{MAM} which contains the max absolute misfit.
}
\description{
Does one step to determine new state values that solves the flow problem
If the model is linear, this one step should suffice. If not-linear it will
try to reduce the misfits by a Newton-Raphson step.
If this improvement is not enough (as e.g. judged
by inspecting the returned misfit statistics), one should
repeat applying this function or use the more general
\code{solve.steps} function
}
\examples{
lingradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M1 = newFLOW2D(domain=domain,systemfluxfunction=lingradflux,name="example")
set.discretisation(M1,nodes=list(type="hexagonal",side=1),method="FV")
add.spatialflux(M1,0.1)
set.BC.fixedstate(M1,"NE",0.1)
solve.step(M1)
nonlingradflux = function(x,y,s,grads) { return(-sqrt(s)*grads)}
M2 = newFLOW2D(domain=domain,systemfluxfunction=nonlingradflux,name="example")
set.isacceptable(M2,function(x,y,s){return(s>=0)})
set.discretisation(M2,nodes=list(type="hexagonal",side=1),method="FV")
add.spatialflux(M2,0.1)
set.BC.fixedstate(M2,"NE",0.1)
solve.step(M2)
}
\seealso{
\code{solve.steps}
}
