% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVFE2D.R
\name{state.fun}
\alias{state.fun}
\title{Get function for current state values}
\usage{
state.fun(model)
}
\arguments{
\item{model}{a two dimensional flow model of the class \code{[A]FElinear2D},
\code{[A]FEquadratic2D}, \code{[A]FEcubic2D},
\code{[A]FEquartic2D} or \code{FV2D} (as generated by \code{set.discretisation})}
}
\value{
a function of two arguments, the space coordinates, that returns
the state at that point of the model as present during constrution.
If the model is of the FV type, the state value of the nearest node is returned.
If the model is of the FE type, a linear interpolation of the states at the nodes
of the triangle that contains the point.
In the FE case also a third parameter eps may be used. If a node exist closer than
eps to the argument, the state value at that node (without interpolation) is returned.
}
\description{
Returns a function which return  state values at points
for the current state of the  model.
}
\examples{
invgradflux = function(x,y,s,grads) { return(-grads)}
domain = matrix(c(1,2,6,1,9,7,4,9),ncol=2,byrow=TRUE)
M = newFLOW2D(domain=domain,systemfluxfunction=invgradflux,name="example")
set.discretisation(M,nodes=list(type="hexagonal",side=1),method="FV")
add.spatialflux(M,0.1)
set.BC.fixedstate(M,"NE",0)
set.BC.fixedstate(M,"SW",0)
solve.steps(M)
sf= state.fun(M)
print(sf(5,5))

}
